Basic support for traffic collection using mitmproxy
Unfortunately, this currently returns the traffic as a string of the
internal mitmproxy flows format, which isn't exactly helpful.

I would prefer to return the traffic as a HAR but mitmproxy only
supports that through a community addon and I'm not sure what the best
way to load that here is.