Fixes #32: Expose mitmproxy events
We only do this for traffic collections on an app analysis.
Implementing it also for system-level traffic collections would have
required an API change in `stopTrafficCollection()`, which would have
made the API worse imo. Also, this is not a stable API anyway (since
we may well switch away from mitmproxy in the future), so it doesn't
matter too much.